Feature 2.0

+++ src/mongoshake/executor/operation.go
@@ -5,14 +5,14 @@ import (
 	"reflect"
 	"strings"
 
-	"mongoshake/dbpool"
-	"mongoshake/oplog"
 	"mongoshake/collector/configure"
 	"mongoshake/common"
+	"mongoshake/dbpool"
+	"mongoshake/oplog"
 
+	LOG "github.com/vinllen/log4go"
 	"github.com/vinllen/mgo"
 	"github.com/vinllen/mgo/bson"
-	LOG "github.com/vinllen/log4go"
 )
 
 var ErrorsShouldSkip = map[int]string{
@@ -21,32 +21,35 @@ var ErrorsShouldSkip = map[int]string{
 
 type CommandOperation struct {
 	concernSyncData bool
+	runOnAdmin      bool // some commands like `renameCollection` need run on admin database
 }
 
 var opsMap = map[string]*CommandOperation{
-	"create":           {concernSyncData: false},
-	"collMod":          {concernSyncData: false},
-	"dropDatabase":     {concernSyncData: false},
-	"drop":             {concernSyncData: false},
-	"deleteIndex":      {concernSyncData: false},
-	"deleteIndexes":    {concernSyncData: false},
-	"dropIndex":        {concernSyncData: false},
-	"dropIndexes":      {concernSyncData: false},
-	"renameCollection": {concernSyncData: false},
-	"convertToCapped":  {concernSyncData: false},
-	"emptycapped":      {concernSyncData: false},
-	"applyOps":         {concernSyncData: true},
+	"create":           {concernSyncData: false, runOnAdmin: false},
+	"collMod":          {concernSyncData: false, runOnAdmin: false},
+	"dropDatabase":     {concernSyncData: false, runOnAdmin: false},
+	"drop":             {concernSyncData: false, runOnAdmin: false},
+	"deleteIndex":      {concernSyncData: false, runOnAdmin: false},
+	"deleteIndexes":    {concernSyncData: false, runOnAdmin: false},
+	"dropIndex":        {concernSyncData: false, runOnAdmin: false},
+	"dropIndexes":      {concernSyncData: false, runOnAdmin: false},
+	"renameCollection": {concernSyncData: false, runOnAdmin: true},
+	"convertToCapped":  {concernSyncData: false, runOnAdmin: false},
+	"emptycapped":      {concernSyncData: false, runOnAdmin: false},
+	"applyOps":         {concernSyncData: true, runOnAdmin: false},
 }
 
 func (exec *Executor) ensureConnection() bool {
 	// reconnect if necessary
 	if exec.session == nil {
-		if conn, err := dbpool.NewMongoConn(exec.MongoUrl, true); err != nil {
+		if conn, err := dbpool.NewMongoConn(exec.MongoUrl, true, true); err != nil {
 			LOG.Critical("Connect to mongo cluster failed. %v", err)
 			return false
 		} else {
 			exec.session = conn.Session
-			exec.bulkInsert = utils.GetAndCompareVersion(exec.session, ThresholdVersion)
+			if exec.bulkInsert, err = utils.GetAndCompareVersion(exec.session, ThresholdVersion); err != nil {
+				LOG.Info("compare version with return[%v], bulkInsert disable", err)
+			}
 		}
 	}
 
@@ -60,6 +63,11 @@ func (exec *Executor) dropConnection() {
 
 func (exec *Executor) execute(group *OplogsGroup) error {
 	count := uint64(len(group.oplogRecords))
+	if count == 0 {
+		// probe
+		return nil
+	}
+
 	lastOne := group.oplogRecords[count-1]
 
 	if conf.Options.ReplayerDurable {
@@ -75,44 +83,45 @@ func (exec *Executor) execute(group *OplogsGroup) error {
 		LOG.Debug("Replay-%d oplog collection ns [%s] with command [%s] batch count %d, metadata %v",
 			exec.batchExecutor.ReplayerId, group.ns, strings.ToUpper(lookupOpName(group.op)), count, metadata)
 
+		/*
+		 * in the former version, we filter DDL here. But in current version, all messages that need filter
+		 * have already removed in the collector(syncer). So here, we only need to write all oplogs.
+		 */
 		// for indexes
-		if conf.Options.ReplayerDMLOnly && hasIndex {
-			// exec.batchExecutor.ReplMetric.AddFilter(uint64(len(group.oplogRecords)))
-		} else {
-			// "0" -> database, "1" -> collection
-			dc := strings.SplitN(group.ns, ".", 2)
-			switch group.op {
-			case "i":
-				err = dbWriter.doInsert(dc[0], dc[1], metadata, group.oplogRecords,
-					conf.Options.ReplayerExecutorInsertOnDupUpdate)
-			case "u":
-				err = dbWriter.doUpdate(dc[0], dc[1], metadata, group.oplogRecords,
-					conf.Options.ReplayerExecutorUpsert)
-			case "d":
-				err = dbWriter.doDelete(dc[0], dc[1], metadata, group.oplogRecords)
-			case "c":
-				err = dbWriter.doCommand(dc[0], metadata, group.oplogRecords)
-			case "n":
-				// exec.batchExecutor.ReplMetric.AddFilter(count)
-			default:
-				LOG.Warn("Unknown type oplogs found. op '%s'", group.op)
-			}
+		// "0" -> database, "1" -> collection
+		dc := strings.SplitN(group.ns, ".", 2)
+		switch group.op {
+		case "i":
+			err = dbWriter.doInsert(dc[0], dc[1], metadata, group.oplogRecords,
+				conf.Options.ReplayerExecutorInsertOnDupUpdate)
+		case "u":
+			err = dbWriter.doUpdate(dc[0], dc[1], metadata, group.oplogRecords,
+				conf.Options.ReplayerExecutorUpsert)
+		case "d":
+			err = dbWriter.doDelete(dc[0], dc[1], metadata, group.oplogRecords)
+		case "c":
+			err = dbWriter.doCommand(dc[0], metadata, group.oplogRecords)
+		case "n":
+			// exec.batchExecutor.ReplMetric.AddFilter(count)
+		default:
+			LOG.Warn("Unknown type oplogs found. op '%s'", group.op)
+		}
 
-			// a few known error we can skip !! such as "ShardKeyNotFound" returned
-			// if mongoshake connected to MongoS
-			if exec.errorIgnore(err) {
-				LOG.Info("Discard known error %v, It's acceptable", err)
-				err = nil
-			}
+		// a few known error we can skip !! such as "ShardKeyNotFound" returned
+		// if mongoshake connected to MongoS
+		if exec.errorIgnore(err) {
+			LOG.Info("Discard known error[%v], It's acceptable", err)
+			err = nil
+		}
 
-			if err != nil {
-				LOG.Critical("Replayer-%d, executor-%d, oplog for [%s] op[%s] failed. (%v) [%v], logs %d. firstLog %v",
-					exec.batchExecutor.ReplayerId, exec.id, group.ns, group.op, reflect.TypeOf(err), err.Error(), count,
-					group.oplogRecords[0].original.partialLog)
-				exec.dropConnection()
+		if err != nil {
+			LOG.Critical("Replayer-%d, executor-%d, oplog for namespace[%s] op[%s] failed. error type[%v]" +
+				" error[%v], logs number[%d], firstLog: %v",
+				exec.batchExecutor.ReplayerId, exec.id, group.ns, group.op, reflect.TypeOf(err), err.Error(), count,
+				group.oplogRecords[0].original.partialLog)
+			exec.dropConnection()
 
-				return err
-			}
+			return err
 		}
 	}
 	// exec.batchExecutor.ReplMetric.ReplStatus.Clear(utils.ReplicaExecBad)
@@ -144,10 +153,16 @@ func buildMetadata(oplog *oplog.PartialLog) bson.M {
 	return bson.M{}
 }
 
-func extraCommandName(o bson.M) (string, bool) {
-	for key := range o {
-		if _, exist := opsMap[key]; exist {
-			return key, true
+func extraCommandName(o bson.D) (string, bool) {
+	//for key := range o {
+	//	if _, exist := opsMap[key]; exist {
+	//		return key, true
+	//	}
+	//}
+
+	for _, ele := range o {
+		if _, exist := opsMap[ele.Name]; exist {
+			return ele.Name, true
 		}
 	}
 
@@ -157,6 +172,13 @@ func extraCommandName(o bson.M) (string, bool) {
 func isSyncDataCommand(operation string) bool {
 	if op, ok := opsMap[strings.TrimSpace(operation)]; ok {
 		return op.concernSyncData
+	}
+	return false
+}
+
+func isRunOnAdminCommand(operation string) bool {
+	if op, ok := opsMap[strings.TrimSpace(operation)]; ok {
+		return op.runOnAdmin
 	}
 	return false
 }
